+++ kameleoon/client/src/Network/NetworkManagerFactory.php
@@ -13,6 +13,7 @@ interface NetworkManagerFactory
         ?string $environment,
         int $defaultTimeout,
         string $kameleoonWorkDir,
-        AccessTokenSourceFactory $accessTokenSourceFactory
+        AccessTokenSourceFactory $accessTokenSourceFactory,
+        ?string $networkDomain
     ): NetworkManager;
 }